Simplified conversion of V9990 cycles into pixel-positions
Main change is to calculate the number of pixels after both the start and end
position are translated to pixel position. I think before this change it was
possible to calculated number of pixels was off by one. Though I don't know of
any test case where this was actually visible.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@11730 7512d703-071c-0410-b5c2-d2bdd895b7d1