Fixed bug in V9990 Bx mode rendering
Occasionally the rendered V9990 image contained lines that had pixels inserted
in random positions (all following pixels on that line are shifted 1 pixel to
the right). This could easily be reproduced by running GEM v0.8 in gfx9000
mode. Thanks to GuyveR800 for reporting this problem.

The problem only occurred with bpp4 or bpp2 modes (the YUV/YJK modes had a
different but related problem). More specifically for bpp4, the problem
occurred when a partial line was drawn that started at an odd pixel (partial
lines are drawn when certain registers are changed in the middle of a
scanline). When this happens we should only use the lower nibble of the first
byte, but the rendering code didn't handle this.

Actually Erik's original V9990 rendering code worked correctly even without
(possibly) skipping the initial nibble in bpp4 mode. To avoid the problem, the
code started rendering one pixel earlier (so that one pixel was renderer
twice). Later when I started to extend the V9990 core with horizontal/vertical
scrolling some of the assumptions were no longer true and the rendering broke.
This was over 5 years ago. It's strange that this problem went unnoticed for so
long. I guess it just indicates how few Gfx9000 software actually exists.

It's also worth mentioning that the corresponding V99x8 rendering code doesn't
suffer from the same problem, while it also doesn't need this complexity of
having to skip nibbles. In the V99x8 code we never render partial lines. Or at
least we always render complete bitmap lines and then only copy the required
portions from that line to the actual render buffer. This allows for much
simpler rendering code, but possibly it renders pixels that will never be
shown. I guess (part of) the reason for this approach is historical: in the
past we had a pre-rendered-bitmap-lines-cache, this cache only worked with
complete lines.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@11731 7512d703-071c-0410-b5c2-d2bdd895b7d1