Detect gz/zip files based on header instead of filename extension
Before this patch we recognized filenames that ended in .gz, .GZ, zip or .ZIP.
Now we check the first 3 or 4 bytes of the content of the file to detect
(g)zipped files.

The main motivation for this change is that we want to choose a different
filename extension for the savestate and replay files. Those files are
compressed XML files. Before this patch the filename was required to end with
.gz, otherwise openMSX wouldn't recognize them (it wouldn't transparently unzip
them).

See also this forum discussion (and mth's reaction):
  http://tasvideos.org/forum/viewtopic.php?t=10272

Note that this patch does not yet change how savestate/replays are saved.
That's something for a follow-up patch.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@11732 7512d703-071c-0410-b5c2-d2bdd895b7d1