Changed some parameters from std::string to const char*
Many methods of XMLElement are very often or even always called with a string
literal as parameter. In many of those cases it's possible to avoid the 'const
char*' -> std::string conversion. This patch enables that.

Made a similar (much smaller) change in CommandLineParser. Combined these makes
the size of the executable about 32kB smaller.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@11736 7512d703-071c-0410-b5c2-d2bdd895b7d1