Try harder to find roms when loading a savestate
When you receive a savestate or replay create by someone else, it often
happens that that file fails to load because it can't find some rom file.

We already tried several options to try to load roms:
 a) The 'resolved' filename. This is the absolute filename of where the rom was
    actually found on the original system.
 b) The filename as originally typed by the user.
 c) Try all the sha1sums as specified in the config file.

'a' is mainly intended for re-use on the same system. For example when you
reload a savestate, but you started openmsx from another working directory.

'b' works fine when you start openmsx with 'openmsx -carta mygame.rom' and you
pass the savestate to another user who also has mygame.rom. So the same
relative path but not necessarily the same absolute path. Unfortunately very
often users specify absolute paths when selecting game roms.

'c' works fine for systemroms, but game roms usually don't have a list of
sha1sums in the config.

This patch adds an extra step (between 'a' and 'b') that searches the (game)rom
based on the sha1sum of the rom. We already stored this sha1sum in the
savestate to be able to check it and give a warning when there was a mismatch.

For the moment this search on sha1sum only works well for systemroms. This is
because we only automatically index the ~/.openMSX/share/systemroms directory.
A next step could be to allow additional user-specified 'rom pools'.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@11741 7512d703-071c-0410-b5c2-d2bdd895b7d1