Fixed bug in CompressedFileAdapter::getModificationDate()
A consequence of this bug was that while walking over the files in a filepool,
compressed files initially got the wrong timestamp in the database. So the next
walk over the files sees a wrong timestamp and recalculates the sha1sum (and
corrects the timestamp). The third walk over the files, won't recalculate the
sha1sum anymore. This is fixed now.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@11742 7512d703-071c-0410-b5c2-d2bdd895b7d1