Speedup filepool walk, don't stat() files twice
During walking over all files in a filepool, each file was stat() twice. Once
to check if it's a file or a directory and later (for files) again to query the
modification date. I don't know if this is measurable in performance, but it
looked a bit silly in the strace output.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@11743 7512d703-071c-0410-b5c2-d2bdd895b7d1