Fix O(N^2) behaviour on walking files in filepool
When walking over all N files in a filepool, we did for each file a O(N) lookup
of the filename. Combined this is a O(N^2) algorithm. This patch fixes it by
making the internal filepool data structure support fast (O(log N)) lookups on
both sha1sum and on filename.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@11744 7512d703-071c-0410-b5c2-d2bdd895b7d1