Optimize combination of File::mmap() and File::getSize()
These two methods were (almost) always used in pairs. And internally mmap()
calls getSize(), so getSize() was effectively called twice. This patch makes
mmap() return the result of getSize().

Before this patch, calling mmap() on a file with filesize 0 was an error. Now
it returns a NULL pointer (and size 0). So now it returns a valid mapping
(though one that cannot be dereferenced, but that's fine) before this threw an
exception.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@11747 7512d703-071c-0410-b5c2-d2bdd895b7d1