Simplified keyboard fix of SVN revision 11324
See commit message of revision 11324 for much more details, but very shortly:
in that patch we added calls to stopReplay() to make sure the arrays
userKeyMatrix and hostKeyMatrix were synchronized before using userKeyMatrix.
In this follow-up patch we drop the synchronization again and instead use
values of the hostKeyMatrix array directly.

This patch is also a preparation for the 'replay viewonly mode' patch Quibus is
working on. In viewonly mode it's possible, it's even the common case, that the
hostKeyMatrix and userKeyMatrix arrays are out-of-sync (in non-viewonly mode,
both arrays are only in exceptional cases out-of-sync). So once the viewonly
mode stuff is committed, this patch is not only a simplification, but also a
correctness fix.

Another reason for committing this patch separately from the viewonly patch is
debug-/testability. Although I currently believe this patch is correct, the
keyboard handling has proven in the past to be quite tricky and a common source
of bugs (like hanging or ignored keys on switches from replay to normal mode).
If we later discover such a bug, it would be nice to be able to test/bisect
this patch separately from the viewonly mode patch.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@11751 7512d703-071c-0410-b5c2-d2bdd895b7d1