Refactor external slot allocation
Before this patch it was not possible to have one extension put two or more
MSXDevices in a single external slot. For example a combination of a ROM and a
RAM. Before this patch these two would each end up in a different external
slot, even though they were grouped in the same <primary slot="any">,
<secondary slot="any"> tags.

This patch only partly solves the above problem. It adds reference counting to
the external slot allocation mechanism. The next patch will make sure that
devices grouped in the same slot-tags also receive the same slot number.

This patch also splits the CartridgeSlotManager methods getSpecificSlot() and
getAnyFreeSlot() into a part to query the slot number and a part to actually
allocate that slot. This makes error handling easier (especially for the
changes in the next patch).

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@11757 7512d703-071c-0410-b5c2-d2bdd895b7d1