Give the same slot to MSXDevices grouped in the same slot-tags.
See message of previous commit for an example.

This is implemented by changing the value of the slot="any" attribute into
the actual value of the primary/secondary slot. So the first device in the
group will search for a free external slot and the next devices in that group
will reuse those values.

We already had a mechanism for MSXDevices to record the actual used slot in the
config file (this is required for savestates, when loading a savestate we must
be sure that devices end up in the same slots). This was done by adding a
<primary_slot> and <secondary_slot> tag in the config section of the MSXDevice
itself. But this mechanism is now no longer needed (we still support it when
loading old savestates).

Because of this change it's now important that extensions with memory devices
have both a <primary slot="any"> and <secondary slot="any"> tag. All our
extensions already had this, except for Philips_VY_0010. Though this was a
mistake, it's fixed now.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@11758 7512d703-071c-0410-b5c2-d2bdd895b7d1