Changed MSXRam and RomPlain MSXDevices
These changes are required for the "MSX-Audio BIOS v1.2" extension (this
extension will be added in the next commit).

MSXRam:
The base address and the size of the RAM are by default derived from the (union
of) the <mem> tag(s) in the device configuration. This patch allows to override
these defaults. This makes it possible to have a (small) RAM that is mirrored
in the Z80 address space.

RomPlain:
There was a check in this device that the combination of (guessed) base address
and size of the ROM fits inside the <mem> specification. This patch disables
this check when the base address of the ROM was explicitly specified. This
allows to for example create a config with a 64kB ROM that is partly hidden by
another memory device.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@11759 7512d703-071c-0410-b5c2-d2bdd895b7d1