Extended MSXCPUInterface::testUnsetExpanded() to also work with MSXMultiMemDevice
testUnsetExpanded() checks whether it's OK to remove a slotexpander. This is
for example not allowed when there are still devices inserted in that expander.
It's also possible to have one extension that is a combination of a slotexpander
and already some devices in that expander. In such a case there can still be
devices in the expander (but only those that are also part of the extension).

testUnsetExpanded() mostly handled this fine. Though it went wrong with
MSXMultiMemDevices. These are used when multiple MSXDevices share the same 16kB
page in a certain (sub)slot. There was actually a TODO about this in the code.

In practice MSXMultiMemDevices are rarely used, and even more rarely in
slotexpanders. But it does occur in the recently added 'Boosted_audio'
extension.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@11765 7512d703-071c-0410-b5c2-d2bdd895b7d1