Fixed bug in (sub)slot handling and savestates
In the following scenario:
 - Start openmsx with
     openmsx -machine msx2 -ext ram1mb
 - In the openmsx console execute
     machine_info issubslotted 1
   this will return '0', slot 1 is not expanded
 - Savestate/loadstate (with or without exiting openmsx)
 - Again perform the same query.
In the last step slot 1 will be expanded. This is a bug.

In some parts of the code we treat an unexpanded slot as subslot 0, in other
parts of the code we want to keep subslot 0 separate from an unexpanded slot.
The bug was caused by mxing up these two cases (after the recent refactoring in
revision 11758).

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@11766 7512d703-071c-0410-b5c2-d2bdd895b7d1