Fixed bug with (new) Japanese DOS2 extension
This extension must be inserted in (any) primary slot. It contains an internal
slotexpander which has a memory mapper in subslot 1 and a ROM mapper in subslot
3. I expect Quibus to commit this extension soon after this bug fix.

The problem was similar to a problem fixed by commit 11757. In a first phase we
would allocate a (free) primary slot and mark that slot as an expanded slot. In
the second phase we would search for a free slot for the two memory devices.
But the two phases don't know about each other, so phase two would allocate
another free slot. Fixed by recording the result of phase 1 in the hardware
config, so that phase 2 doesn't need to search again.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@11769 7512d703-071c-0410-b5c2-d2bdd895b7d1