On loadstate also search diskimages by sha1sum
The current implementation prefers the original file (if it (still) exists on
the current host system) over a file with an exact sha1sum. Thus it only
searches in the filepool when the original file doesn't exist. It does give a
warning when the sha1sum does not match (anymore).

An alternative is to always search the filepool first (or only when the
origiinal file's sha1sum has changed) before using the original file.

In practice, very often, the difference between these alternatives doesn't
matter: how often do you change a disk image while you still have a copy
of the original. Though in the (rare) case this does happen I thought the user
may have intentionally copied and modified the file, so it might be better
to keep using that copy in the savestate.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@11786 7512d703-071c-0410-b5c2-d2bdd895b7d1