Extended IFF flags in "CPU regs" debuggable
Added an extra flag 'IFF1 AND last-instruction-was-not-EI', so this indicates
whether the CPU could potentially accept an interrupt at the start of the
current instruction.

The main motivation to add this flag is the following: we want to detect bugs
in a MSX program where we write to the VDP regiters, but forget to disable
the interrupts. At first sight this seems already possible with something
like this:
	debug set_watchpoint write_io 0x99 {[reg iff] & 1}
So on every write to IO port 0x99 we check that IFF1 flag is set, and if so we
break. (The IFF1 flag corresponds to EI/DI mode of the CPU.)

But this triggers false positives, for example in the MSX-BIOS:
	1804  di
	1805  out (#99),a
	1807  ld  a,#92
	1809  ei
	180A  out (#99),a
This code is legal because the CPU still doesn't accept interrupts for one
instruction following an EI instruction.

After this patch we can change the condition on the watchpoint to
	{[reg iff] & 4}
to get the correct behaviour.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@11792 7512d703-071c-0410-b5c2-d2bdd895b7d1