Recognize optional <window> tag inside <rom> in hardwareconfig.xml
For example:
	<ROM id="KANJI BASIC ROM">
	  <mem base="0x4000" size="0x04000"/>
	  <rom>
	    <sha1>42f4e3360eb134857b0efa82c1abedf342680124</sha1>
	    <filename>roms/ascii_msxdos22.rom</filename>
	    <window base="0xC000" size="0x4000"/>
	  </rom>
	</ROM>

This creates a 16kB ROM with content the last 16kB of the 64kB file
'ascii_msxdos22.rom'. This functionality is required to implement the 'ASCII
MSX-DOS2 cartridge'. This cartridge has an internal slotexpander with memory
mapper, DOS2 rom mapper and a 16kb Kanji rom which is a mirror of the last 16kB
block of the dos2 rom.

In the past we had a similar functionality provided by the <skip_headerbytes>
and <filesize> tags. Support for these two tags was removed in revision 11568
because (at the time) this functionality was not used and it was buggy when
used in combination with the <sha1> tag. Instead of re-enabling these two tags
(and fixing the sha1 related bug) I also choose a different syntax. Mainly
because the old names were not very clear: in the example above we don't really
skip a header and the name 'filesize' is wrong (or at least confusing) in this
context.

In a turbor machine config we also already have functionality that's similar to
this new <window> feature, namely the <firstblock> and <lastblock> tags. I
think it's possible to replace first/lastblock with window, though that's
slightly less efficient (the current implementation of window requires to mmap
the big rom file multiple times in memory instead of sharing it as done with
first/lastblock). And we anyway have to keep first/lastblock for backwards
compatibility.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@11793 7512d703-071c-0410-b5c2-d2bdd895b7d1