Simplified filepool command, removed priority
Removed the '-priority' property from filepool entries. Instead the priority is
now implicit in the order of the entries. It's possible to add new entries in a
specific position (by default new entries are added at the end).

Before this patch 'filepool list' already sorted entries according to ascending
priority, so in a way the order was already the priority.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@11796 7512d703-071c-0410-b5c2-d2bdd895b7d1