Refactored GlobalCliComm and CliConnection
Split CliConnection into a CliListener base class and a subclass that does the
transformation to XML and the actual communication with external clients. Also
simplified GlobalCliComm: removed all XML formatting and the special case for
stdout message output. That stdout output is now implemented as another
CliListener.

Though the main motivation for this patch is new 'message_display_callback'
stuff (will be implemented in one of the next patches). This new settings will
allow to pass CliComm::log() messages to Tcl (where we can show them using
OSD).

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@11799 7512d703-071c-0410-b5c2-d2bdd895b7d1