Added TclCallbackMessages
This makes 'cli message' available to a Tcl callback proc. For example:
	set message_callback osd::menu::display_osd_text

We still need to polish this further .. not all messages are useful (e.g.
message that a png is written when a savestate is created). The original reason
for working on this is to show a message while scanning a filepool (because it
can take a long time), that's also still TODO.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@11800 7512d703-071c-0410-b5c2-d2bdd895b7d1