Fixed link error when using (ubuntu?) gcc-4.5
We use the function clock_gettime() which (when using glibc) is defined in the
library librt. So we should pass the option -lrt to the linker.

For some reason often this library was already included via some other way, but
I got a link error when I switched to the ubuntu gcc-4.5 compiler. I did not
see this error in the past with my own build of gcc-4.5 nor with my current
build of gcc-4.6. But since we use librt anyway it's better to explicitly add
it.

Note that other libc distributions may put the clock_gettime() function in
another library (e.g. uClibc puts it in libc itself). Ideally we should test
for this, though this patch hardcodes this knowledge.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@11812 7512d703-071c-0410-b5c2-d2bdd895b7d1