Added text-wrapping to OSD text widgets
Before this patch OSD text widgets had no maximum width. Thus if there are no
newline characters inside the text string, everything is rendered on a single
line (support for newline chars was also only added in the previous patch). Now
it's possible to specify a maximum width for the widget and the text will be
split in multiple lines so that it fits in that specified width.

We added 3 new properties on text widgest:
 -wrap: specifies the wrap mode, possible values are
    none: (default) Same as old behaviour, next two properties are ignored.
    char: Split text at character boundaries.
    word: Split text at word boundaries. A word boundary means right after a
          '-', '/' or ' ' character. Also trailing spaces on the current and
          leading spaces on the next line are removed.
 -wrapw: (default=0.0) specify maximum text width in pixels (possibly scaled)
 -wraprelw: (default=1.0) maximum width as a fraction of the parent widget
            width

The -wrapw and -wraprelw can be combined and can even be negative. For example:

   osd create text my-text-widget -x 2 - y 2 -wraprelw 1.0 -wrapw -4 \
                                  -size 20 -text "My possibly wrapped text."

This will create a text widget that is 4 pixel less wide than the parent widget
width a border of 2 pixels and the left and right.

TODO multi-byte utf8 characters are not yet handled correctly, will be fixed
in the next patch.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@11815 7512d703-071c-0410-b5c2-d2bdd895b7d1