Added -query-size property on osd text widgets
Before this patch it was not possible to draw a rectangle around some text
because you couldn't know the rendered height of the text (you don't know
how many lines there are after wrapping). This patch adds a read-only
-query-size property on text widgets. this property returns the width and
height of the bounding box of the rendered text.

This is implemented by rendering the text (if not done so already) and then
get the width and height of the internal image that represents that rendered
text. To do this we had to refactor the code so that the rendering of that
internal image can be done without an actual output-surface (that surface
was anyway only used to query the width and height of the output window).

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@11817 7512d703-071c-0410-b5c2-d2bdd895b7d1