Fixed crash when the default OSD font could not be loaded
When the default OSD font skins/Vera.ttf.gz could not be loaded, openMSX
crashed. This can happen either because the file itself is missing (in both
openMSX system and user directories) or because the file could not be unzipped
in /tmp (e.g. partition is full).

Fixing this crash was quite complex, it actually required 3 fixes. For each of
these 3, the problem starts by trying to print a message on OSD using the new
Tcl CliComm-callback mechanism. But when the OSD font is missing, this triggers
the printing of another error message:

* First problem was a deadlock in GlobalCliComm caused by trying to acquire
the semaphore in GlobalCliComm::log() twice. Fixed by not delivering recursive
log messages (the 2nd recursive message is only printed on stderr). See comment
in GlobalCliComm::log() for more details.

* Second problem was caused by iterating over a changing vector of OSDWidgets
or still using an already deleted OSDWidget. Printing an OSD message both
deletes the old OSD text box and creates a new box, but if this callback is
triggered by an error from the OSD drawing (thus when we're iterating over the
widgets), this causes the above problems. Fixed by changing the type of the
OSDWidget::SubWidgets collection to a vector of shared_ptr (was raw pointers)
and by making a copy of this vector before iterating over it in the draw loop.

* After these two fixes, the crash was gone, but openMSX was still flooded with
error messages in a tight loop. Not a recursive loop that causes a
stack-overflow, but one message for each rendered frame: in the message
callback we create the OSD widgets, but only in the next frame when the widgets
are drawn, the next error is generated. To break this loop, I added a special
property -suppressErrors, when this property is set, we don't generate errors
for that (subtree) of OSD widgets.

Even after these fixes we still get periodic attempts to show an OSD error
message. This is caused by the time-indication-text in the reversebar which is
changing every second (even though the reversebar is mostly hidden). The
message is visible as a small orange rectangle at the top of the screen, (of
course) this rectangle contains no text. On stdout the error messages are
readable.

This is not ideal, but I don't see something (easy) we can do about this. We
can't show an error inside the openMSX window, because the error is caused in
the first place by not being able to show text. We could make this a fatal
error (refuse to start openMSX), but that won't give any visual feedback to GUI
users who don't usually look at stdout/err. We could still try to remove the
empty orange rectangle caused by the reversebar stuff, but that will also
remove any visual clue about why other OSD stuff possibly also doesn't work as
it should. Also I don't believe this will be a common situation. When this
happens, there's mostly something wrong with the openMSX installation.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@11831 7512d703-071c-0410-b5c2-d2bdd895b7d1