Fixes in UnicodeKeymap
I got this error while starting openMSX after a SVN update:
    ng deadkey number in keymap file. It must be 1..

There was a bug in the construction of this string, because the intended
message was:
    Wrong deadkey number in keymap file. It must be 1..3

In C++ operator+(const char*, int) does not always what you expect ;-)

I got this error because the format of the unicodemap file has changed. The
keyword DEADKEY is replaced with DEADKEY1. All unicodemap files in SVN were
updated accordingly. But I didn't install these new files yet. It's possible
for other users as well to run into this problem. So I extended the parser to
also still recognize the old keyword.

Also cleaned up the code a bit and fixed a small memory leak (still have to
review the recent Keyboard related changes in more detail).

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@11864 7512d703-071c-0410-b5c2-d2bdd895b7d1