Fixed bug in sprite collision detection
Sprites cannot collide in the (left or right) border. See comments in the code
for more details.

Thanks to n_n for providing the initial testcase that lead to the discovery of
this bug. Thanks to mth for reviewing and suggesting a more elegant way to
handle the left border case.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@11868 7512d703-071c-0410-b5c2-d2bdd895b7d1