In fast-forward mode don't trigger break/watchpoints or conditions
When using reverse we emulate in fast-forward mode (starting from a nearby
snapshot). However when during fast-forward we hit a breakpoint openMSX would
hang because we're emulating in a tight loop till we reach a certain moment in
time, but we're not making progress because CPU emulation immediately returns
without advancing in time.

Solved by not triggering breakpoints, watchpoints or debug conditions while in
fast-forward mode. As a bonus fast-forward emulation will now go faster in case
there are active breakpoints.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@11875 7512d703-071c-0410-b5c2-d2bdd895b7d1