Fixed calculation of 'startcluster'
We forgot to multiply by SECTOR_SIZE, so effectively all FAT sectors were
treated the same as the first FAT sector, so once the disk was half-full
(some) things could start going wrong. This bug was alreadt present in the
very first version of DirAsDSK.

(Though I think this is an important bug, it was not the cause of the problem
that zett reported recently. See next commit.)

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@11887 7512d703-071c-0410-b5c2-d2bdd895b7d1