Fixed disk corruption with DirAsDSK
See comments in DirAsDSK::readSectorImpl() for a detailed explanation of this
problem. Very short summary: problem was triggered by the extra periodic sector
reads that are triggered by the creation of the reverse snapshots (to calculate
a sha1sum for the disk image). These reads came in between sector writes at a
point where the DirAsDSK algorithms can't handle them. Solved by doing these
extra sector reads in 'peek-mode'. This is not ideal because in this mode we
don't sync with the host filesystem and thus the sha1sum is not guaranteed to
be up-to-date. Though I can't think of a better solution right now, and it's
certainly better than disk corruption.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@11888 7512d703-071c-0410-b5c2-d2bdd895b7d1