Reverted part of commit 11895
Using std::map::operator[] or map::insert() in insert elements in a map is NOT
equivalent:

	myMap[myKey] = myValue;

This syntax requires that the key and value types can be default constructed
(that's what revision 11895 changed). It will _overwrite_ the value for an
existing key in the map.

	myMap.insert(make_pair(myKey, myValue);

This will NOT overwrite an already exiting key.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@11905 7512d703-071c-0410-b5c2-d2bdd895b7d1