Optimize displaying of ram_watch / allow changing an existing ram_watch address
Unlike arrays, dicts have a stable order. Use this to avoid having to sort the
addresses each time they are displayed. This also allows to more efficiently
iterate over the addresses (using 'dict for').

Before it was an error to set a ram_watch on an address that was already
watched. Now it's not an error anymore. This can bw used to change the
properties of an already set ram_watch.

It's also not an error anymore to remove a not-existing ram_watch or to use
'ram_watch save' when no ram_watches were defined.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@11914 7512d703-071c-0410-b5c2-d2bdd895b7d1