Fixed message_callback related error when closing openMSX during recording
When openMSX was quit during a video-recording:
  record start
  exit
You got a message like this:
  Error executing message_callback function "osd::display_message":
  invalid command name "after"

This was because during the shutdown sequence, the after command was already
destroyed (unregistered) while a little later the Tcl message_callback function
was triggered which tried to use the 'after' command. Fixed by both moving the
de-registration of the message_callback earlier and de-registration of the
after command later in the shutdown sequence.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@11918 7512d703-071c-0410-b5c2-d2bdd895b7d1