Disable MLAA scaler for the upcoming 0.8.1 release
This scaler is not finished yet (e.g. screen7 with scale_factor=3, laserdisc,
openGL). Maarten is also still tweaking the actual algorithm. So we decided to
disable this scaler for the release.

I only removed the value 'mlaa' from the possible scaler_algorithm values. The
actual mlaa code is still included in the binary. I checked that it's only
about 20kB of code, so this shouldn't matter that much.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@11936 7512d703-071c-0410-b5c2-d2bdd895b7d1