Fix crash on exit when there were (conditional) breakpoints defined
To reproduce:
 - start openmsx
 - run the 'text_echo' command      (or set a breakpoint with a condition)
 - exit openmsx

Before this patch this would result in a crash with message
  called Tcl_FindHashEntry on deleted table

The reason is that breakpoints (and conditions) are kept in a global map.
This map is only cleaned up after main() has been exited. But by that time
the Tcl interpreter has already been destoyed. (Destroying conditional
breakpoints still deletes some TclObjects.)
Fixed now by adding an explicit cleanup of these global structures. Though
it would be nicer to replace them with some non-global structures.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@11963 7512d703-071c-0410-b5c2-d2bdd895b7d1