Swapped order of cmdEngine->sync() and renderer->updateXXX()
Before this patch I could trigger this scenario:

VDPVRAM::updateDisplayEnabled()
{
	...
	renderer->updateDisplayEnabled();
	   --> PixelRenderer::updateDisplayEnabled()
	    --> PixelRenderer::sync()
	     --> PixelRenderer::renderUntil()
	cmdEngine->sync();
	   --> VDPCmdEngine::sync()
	    --> HmmmCommand::execute()
	     --> VDPVRAM::cmdWrite()
	      --> VDPVRAM::writeCommon()
	       --> VRAMWindow::notify()
	        --> PixelRenderer::updateVRAM()
	         --> PixelRenderer::renderUntil()
	...
}

So in VDPVRAM::updateDisplayEnabled(), the first statement
renderer->updateDisplayEnabled() indirectly triggers
PixelRenderer::renderUntil(), this is done for timestamp 'now'. The second
statement cmdEngine->sync() also triggers PixelRenderer::renderUntil(), but
this is possibly done for timestamps between 'last-cmdEngine-sync' and 'now'.
So from the standpoint of PixelRenderer::renderUntil(), time appears to go
backwards.

Simply swapping the order of the two statements above fixes the problem. There
is a similar situation in updateSpritesEnabled() and updateDisplayMode().

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@11974 7512d703-071c-0410-b5c2-d2bdd895b7d1