Disable OSD mouse coordinates when host mouse cursor is not visible.
The host mouse cursor is hidden in fullscreen mode, in grabinput mode, but also
after some delay depending on the 'pointer_hide_delay' setting.

The reason for doing this is that otherwise (e.g. when using the mouse in an
MSX program) it's possible to accidentally click on the reverse bar. This will
also block the OSD mouse in other Tcl scripts (e.g. vampier's nemesis script),
but almost always those scripts will also not be useful when the host mouse
cursor is not visible.

I also changed the default value for the 'pointer_hide_delay' setting from 1.0
to 2.0. Otherwise it sometimes happens when you want to click in the reverse
bar, that the host cursor has already disappeared. For example when you want to
go to a specifc point in time, and you're reading those yellow popups in the
reverse bar. With a delay of 2 seconds it works comfortably for me.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@11987 7512d703-071c-0410-b5c2-d2bdd895b7d1