Revert "Deserialize some VRAMWindow objects based on VDP state"
This reverts commit 11973. That commit de-serialized some VRAMWindows based on
other VDP state. It was done as preparartion for another refactoring, but in
the end it turned out to not be necessary for that refactoring.

There was also an unexpected problem in this patch: restoring (some) VRAMWindow
objects was done via VDP::resetMasks(). Normally calling this method is fine
for this purpose. Though during de-serialization the VDP subsystem is not yet
completely in a consistent state, and this causes problem indirectly via the
updateWindow() notifications.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@11992 7512d703-071c-0410-b5c2-d2bdd895b7d1