Fixed OSD mouse coordinates on dingoo
This bug caused the osd_widgets::text_box to not fade out anymore. This happens
when the mouse is located above the widget. When the host mouse cursor is
hidden we return (NaN, NaN) for the mouse coordinates. For some reason on
dingoo this location is (wrongly) thought to be inside the text_box widget. See
comments in OSDWidget.cc for more details. Fixed by returning the coordinates
(+inf, +inf) instead.

git-svn-id: file:///openmsx-svn/code/openmsx/tags/RELEASE_0_8_1@12003 7512d703-071c-0410-b5c2-d2bdd895b7d1