Fixed crash on startup
To reproduce:
 - create a file 'run.tcl' with as (only) content
      ext gfx9000
 - start openmsx with this script
      openmsx -script run.tcl

Early during startup, openMSX always starts with the dummy renderer active.
Later we switch to the active renderer. Switching renderer uses the openmsx
event mechanism (see comments in Display.cc for the reason why). In the
scenario above, we already restored the renderer setting, but the event that
does the actual switching was not yet delivered. This confused the video
subsystem and resulted in a crash. Fixed by forcing event delivery right after
restoring the renderer setting.

Bug was discovered by nerlaska and reported by Vampier.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12020 7512d703-071c-0410-b5c2-d2bdd895b7d1