Use Tcl-8.5 'in' and 'ni' operators to replace lsearch
Tcl-8.5 has new operators 'in' and 'ni' to respectively test if an element is
'in' or 'not-in' a given list. So we can simplify the following code
	if {[lsearch -exact $my_list $my_element] != -1} {
to
	if {$my_element in $my_list} {
And similarly for the 'ni' operator.

Actually before this patch we had potential bugs in our scripts because we
often forgot to add the -exact flag to the lsearch command (without this flag
lsearch interprets the last parameter as a glob pattern).

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12021 7512d703-071c-0410-b5c2-d2bdd895b7d1