Prefer to use 'eq' and 'ne' for string comparison in Tcl
Before this patch we always used the operators == and != in Tcl scripts to
compare either numbers or strings (or sometimes other types). Though for all
but numeric comparisons the 'eq' (equal) and 'ne' (not-equal) operators are
often better.

Operators '==' and '!=' will first try to convert both rhs and lhs to integer
or float and then numerically compare both sides. If either of these (or both)
conversions fail then lhs and rhs are compared as strings (at least
conceptually, internally it might be implemented more efficiently). Operators
'eq' and 'ne' always use string comparison.

For example:
  1 == 0x1      is true
  1 eq 0x1      is false
  1 == 1.0      is true
  1 eq 1.0      is false

So for string comparison always use 'eq' and 'ne':
 - In rare cases (see examples above) '==' and '!=' will give the wrong result
 - If you know it's not a numeric value, it's a bit faster because Tcl doesn't
   attempt to convert to integer/float first.
For numeric comparison always use '==' and '!=':
 - It allows to mix different representations of the same number (e.g. hex)
 - If you know it's a valid numeric, it's usually faster than string comparison

(Note: this is NOT a new feature in Tcl-8.5)

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12028 7512d703-071c-0410-b5c2-d2bdd895b7d1