Use Tcl-8.5 argument expansion '{*}' to avoid 'eval'
Tcl-8.5 has a new syntax to 'expand' a list and pass it as individual arguments
to some other proc (instead of passing 1 argument which is a list).

For example:
  set l [list "foo" "bar"]
  my_proc $l        <- pass 1 argument to my_proc (which is a list of 2 elements)
  my_proc {*}$l     <- pass 2 arguments to my_proc
  eval "my_proc $l" <- also pass 2 arguments (for Tcl-8.4)

This can be used in some places to avoid having to use 'eval'. This often results
in a simpler syntax (e.g. requiring less escaping).

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12030 7512d703-071c-0410-b5c2-d2bdd895b7d1